requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

pydantic
pydantic-settings

python-dotenv

passlib[bcrypt]
python-jose[cryptography]

email-validator
